[Refactor] #1796 Moved monster-loader.cpp/h from load/ to load/old/ (これらのファイルはまだ現役)
