+++ Hengband/Hengband/Hengband.vcxproj
@@ -500,7 +500,7 @@
     <ClCompile Include="..\..\src\load\old\load-v1-7-0.cpp" />
     <ClCompile Include="..\..\src\load\load-zangband.cpp" />
     <ClCompile Include="..\..\src\load\lore-loader.cpp" />
-    <ClCompile Include="..\..\src\load\monster-loader.cpp" />
+    <ClCompile Include="..\..\src\load\old\monster-loader-savefile10.cpp" />
     <ClCompile Include="..\..\src\load\option-loader.cpp" />
     <ClCompile Include="..\..\src\load\player-attack-loader.cpp" />
     <ClCompile Include="..\..\src\load\player-info-loader.cpp" />
@@ -922,6 +922,7 @@
     <ClInclude Include="..\..\src\action\open-close-execution.h" />
     <ClInclude Include="..\..\src\action\open-util.h" />
     <ClInclude Include="..\..\src\action\run-execution.h" />
+    <ClInclude Include="..\..\src\load\old\savedata10-monster-flag-types.h" />
     <ClInclude Include="..\..\src\load\player-class-specific-data-loader.h" />
     <ClInclude Include="..\..\src\load\savedata-old-flag-types.h" />
     <ClInclude Include="..\..\src\monster-race\monster-aura-types.h" />
@@ -1245,12 +1246,12 @@
     <ClInclude Include="..\..\src\load\old\load-v1-5-0.h" />
     <ClInclude Include="..\..\src\load\load-zangband.h" />
     <ClInclude Include="..\..\src\load\lore-loader.h" />
-    <ClInclude Include="..\..\src\load\monster-loader.h" />
+    <ClInclude Include="..\..\src\load\old\monster-loader-savefile10.h" />
     <ClInclude Include="..\..\src\load\old-feature-types.h" />
     <ClInclude Include="..\..\src\load\option-loader.h" />
     <ClInclude Include="..\..\src\load\player-attack-loader.h" />
     <ClInclude Include="..\..\src\load\player-info-loader.h" />
-    <ClInclude Include="..\..\src\load\savedata-flag-types.h" />
+    <ClInclude Include="..\..\src\load\old\savedata10-item-flag-types.h" />
     <ClInclude Include="..\..\src\load\store-loader.h" />
     <ClInclude Include="..\..\src\load\world-loader.h" />
     <ClInclude Include="..\..\src\racial\racial-util.h" />
