+++ Hengband/Hengband/Hengband.vcxproj.filters
@@ -1457,9 +1457,6 @@
     <ClCompile Include="..\..\src\load\load-util.cpp">
       <Filter>load</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\load\monster-loader.cpp">
-      <Filter>load</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\load\load-zangband.cpp">
       <Filter>load</Filter>
     </ClCompile>
@@ -2362,6 +2359,9 @@
       <Filter>load\old</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\load\old\item-loader-savefile10.cpp">
+      <Filter>load\old</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\load\old\monster-loader-savefile10.cpp">
       <Filter>load\old</Filter>
     </ClCompile>
   </ItemGroup>
@@ -4070,12 +4070,6 @@
     <ClInclude Include="..\..\src\load\load-util.h">
       <Filter>load</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\load\savedata-flag-types.h">
-      <Filter>load</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\src\load\monster-loader.h">
-      <Filter>load</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\src\load\load-zangband.h">
       <Filter>load</Filter>
     </ClInclude>
@@ -5119,6 +5113,15 @@
       <Filter>load\old</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\load\old\item-loader-savefile10.h">
+      <Filter>load\old</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\load\old\monster-loader-savefile10.h">
+      <Filter>load\old</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\load\old\savedata10-item-flag-types.h">
+      <Filter>load\old</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\load\old\savedata10-monster-flag-types.h">
       <Filter>load\old</Filter>
     </ClInclude>
   </ItemGroup>
