+++ src/load/floor-loader.cpp
@@ -9,10 +9,10 @@
 #include "io/uid-checker.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
-#include "load/monster-loader.h"
 #include "load/old-feature-types.h"
 #include "load/old/item-loader-savefile10.h"
 #include "load/old/load-v1-5-0.h"
+#include "load/old/monster-loader-savefile10.h"
 #include "monster-race/monster-race.h"
 #include "monster/monster-info.h"
 #include "monster/monster-list.h"
