+++ src/load/old/load-v1-5-0.cpp
@@ -14,9 +14,9 @@
 #include "grid/trap.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
-#include "load/monster-loader.h"
 #include "load/old-feature-types.h"
 #include "load/old/item-loader-savefile10.h"
+#include "load/old/monster-loader-savefile10.h"
 #include "mind/mind-weaponsmith.h"
 #include "monster-floor/monster-move.h"
 #include "monster-race/monster-race.h"
