+++ src/load/old/monster-loader-savefile10.cpp
@@ -1,8 +1,8 @@
-﻿#include "load/monster-loader.h"
+﻿#include "load/old/monster-loader-savefile10.h"
 #include "load/angband-version-comparer.h"
 #include "load/load-util.h"
 #include "load/old/load-v1-5-0.h"
-#include "load/savedata-flag-types.h"
+#include "load/old/savedata10-monster-flag-types.h"
 #include "system/monster-type-definition.h"
 #include "system/player-type-definition.h"
 #include "util/bit-flags-calculator.h"
