+++ src/load/old/monster-loader-savefile10.h
