+++ src/load/old/savedata10-item-flag-types.h
@@ -32,22 +32,3 @@ enum class SaveDataItemFlagType {
 	ART_FLAGS = 0x10000000,
 	STACK_IDX = 0x20000000,
 };
-
-enum class SaveDataMonsterFlagType {
-	AP_R_IDX = 0x00000001,
-	SUB_ALIGN = 0x00000002,
-	CSLEEP = 0x00000004,
-	FAST = 0x00000008,
-	SLOW = 0x00000010,
-	STUNNED = 0x00000020,
-	CONFUSED = 0x00000040,
-	MONFEAR = 0x00000080,
-	TARGET_Y = 0x00000100,
-	TARGET_X = 0x00000200,
-	INVULNER = 0x00000400,
-	SMART = 0x00000800,
-	EXP = 0x00001000,
-	MFLAG2 = 0x00002000,
-	NICKNAME = 0x00004000,
-	PARENT = 0x00008000,
-};
